Detecting Structural Change with Heteroskedasticity

Calculates the sup MZ value to detect the unknown structural break points under Heteroskedasticity

supmz(formula, data, nBoot = 100)

# S3 method for default
supmz(formula, data, nBoot = 100)

Arguments

formula
Formula for the linear model to be used. It may contain any number of independent variables.

data
Data frame containing dependent and independent variables.

nBoot
Number of bootstrap samples to compute the critical region.

Value

MZ Gives values of MZ as given by Mumtaz et.al (2017)

BreakLocation Provides the data point position where the structural break occured

SupMzValue Returns the supremum value from MZ values

SupMZ0 Returns the bootstrapped critical value for testing the significance of SupMZ

nBoot Shows the number of bootstrap samples used to compute the critical region

References

Mumtaz Ahmed, Gulfam Haider & Asad Zaman (2017). Detecting structural change with heteroskedasticity. Communications in Statistics - Theory and Methods. 46(21):10446-10455, DOI: 10.1080/03610926.2016.1235200

Examples

data(Japan)
fm1 <- supmz(formula = C~Y, data = Japan, nBoot = 10)
fm1

#> $MZ
#>  [1]       NA       NA       NA       NA       NA 40.13318 41.96136 44.75228
#>  [9] 47.53959 50.50748 54.63145 58.35925 62.02521 64.42628 68.17842 69.47337
#> [17] 65.97529 65.14546 67.66180 68.56435 69.83502 69.16709 60.10518 64.66533
#> [25] 57.69157 51.48075
#> 
#> $BreakLocation
#> [1] 21
#> 
#> $SupMzValue
#> [1] 69.83502
#> 
#> $SupMZ0
#>      95% 
#> 42.50472 
#> 
#> $nBoot
#> [1] 10
#> 

data(Belgium)
fm2 <- supmz(formula = C~Y, data = Belgium, nBoot = 10)
fm2

#> $MZ
#>  [1]       NA       NA       NA       NA       NA 37.83128 45.25909 46.32898
#>  [9] 49.23442 48.69960 51.02426 51.94672 38.81871 42.84383 43.10587 41.49724
#> [17] 40.27792 40.83222 39.74601 39.07733 34.95951 31.81879 29.46918 29.30836
#> [25] 27.54211 24.22064
#> 
#> $BreakLocation
#> [1] 12
#> 
#> $SupMzValue
#> [1] 51.94672
#> 
#> $SupMZ0
#>      95% 
#> 28.33269 
#> 
#> $nBoot
#> [1] 10
#> 

data(Srilanka)
fm3 <- supmz(formula = C~Y, data = Srilanka, nBoot = 10)
fm3

#> $MZ
#>  [1]       NA       NA       NA       NA       NA 55.33641 56.86234 52.93426
#>  [9] 56.04649 59.87044 64.18047 68.11976 72.89355 76.70137 81.16258 83.30058
#> [17] 85.23767 89.87243 85.16271 89.53792 84.39359 52.38645 55.09757 59.14269
#> [25] 62.56804
#> 
#> $BreakLocation
#> [1] 18
#> 
#> $SupMzValue
#> [1] 89.87243
#> 
#> $SupMZ0
#>      95% 
#> 49.92042 
#> 
#> $nBoot
#> [1] 10
#>